Move PC aux action and ammo retrieval to helpers
